refactor: 바로팟 생성 페이지 리팩토링
- MetaData 추가 및 바로팟 생성 페이지 리팩토링
