--- src/app/(afterLogin)/restaurants/[restaurantId]/baropot/[baropotId]/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function page() {
-  return <div>page</div>;
-}
